requirements.txt
openenv-core
openai
fastapi
uvicorn
pydantic
python-dotenv